created basic set up for header and footer
