+++ package.json
@@ -6,7 +6,7 @@
     "client-dev": "parcel public/index.html --open",
     "server-dev": "nodemon server.js",
     "seed": "node server/seed.js",
-    "start": "parcel build public/index.html && npm run seed && node server.js"
+    "start": "parcel build public/index.html --open && npm run seed && node server.js"
   },
   "dependencies": {
     "@babel/preset-env": "^7.15.0",
