+++ public/react/components/Footer.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Footer = () => {
+  return (
+    <footer style={{ height: "8vh", backgroundColor: "blue" }}>
+      <h2>this is footer</h2>
+    </footer>
+  );
+};
+
+export default Footer;
