+++ public/react/components/Header.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Header = () => {
+  return (
+    <div>
+      <h2>InvetoryApp</h2>
+    </div>
+  );
+};
+
+export default Header;
